backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==4.0.1
transformers==4.9.2
torch==1.9.0
numpy==1.21.2